brain/requirements.txt
torch==2.0.1
kafka-python==2.0.2
numpy==1.24.3
